refactor: move resume and dependency ordering into monochange_publish
